split into analyzer & react component
